propose clarifying name change
libtonode tests named

finish removing ambiguous 'integration'

reorg chain generics to make 'cargo nextest run -p libtonode chain_generics' work

this makes chain_generic invocation a little cleaner, at the cost of a very minor conflict
