+++ .github/workflows/coverage.yaml
@@ -19,7 +19,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Symlink lightwalletd and zcash binaries
-        run: ln -s /usr/bin/lightwalletd /usr/bin/zcashd /usr/bin/zcash-cli ./integration-tests/regtest/bin/
+        run: ln -s /usr/bin/lightwalletd /usr/bin/zcashd /usr/bin/zcash-cli ./libtonode-tests/regtest/bin/
 
       - name: Symlink zcash parameters
         run: ln -s /root/.zcash-params /github/home
