+++ docs/libtonode-tests.txt
@@ -1,8 +1,8 @@
 1. follow the instructions in ./TEST-REGTEST.md in order to set up the local regtest server, with the caveat that
-the root directory is integration-tests/regtest NOT zingocli/regtest, so in particular the binaries will look like this:
+the root directory is libtonode-tests/regtest NOT zingocli/regtest, so in particular the binaries will look like this:
 
 
-integration-tests/regtest/bin/
+libtonode-tests/regtest/bin/
 ├── lightwalletd
 ├── zcash-cli
 └── zcashd
