+++ libtonode-tests/tests/data/wallets/missing_data_test/zingo-wallet.dat
