+++ libtonode-tests/tests/data/wallets/v26/202302_release/mainnet/zingo-wallet.dat
