+++ libtonode-tests/tests/data/wallets/v26/202302_release/regtest/orch_and_sapl/zingo-wallet.dat
