+++ libtonode-tests/tests/shield_transparent.rs
