+++ libtonode-tests/tests/zingo-wallet-v26-2.dat
