+++ libtonode-tests/tests/zingo-wallet-v26.dat
